Added material-ui appbar
Have been using a mix of an old version of material-ui, react-mdl, and the newest version of material-ui due to following a online tutorial. Now, we are trying to migrate to make all frontend component to the newest version of material-ui. This commit is starting off by adding a material-ui appbar.